requirements.txt
streamlit
sentence-transformers
faiss-cpu
py2neo
pypdf2
numpy
pdfplumber
neo4j
